[M-plane] Optionally use IQ bitwidth from the config file
* use uint8_t instead of int16_t for `iq_width` (the value cannot be
  negative)
* move <compression-type> outside the if condition since it's a mandatory
node
* use <compression-method> node

+++ radio/fhi_72/mplane/ru-mplane-api.h
@@ -102,7 +102,7 @@ typedef struct {
 typedef struct {
   char *ru_mac_addr;
   uint32_t mtu;
-  int16_t iq_width;
+  uint8_t iq_width;
   uint8_t prach_offset;
 
   // DU sends to RU and xran
